+++ packages/vite/src/node/plugins/workerImportMetaUrl.ts
@@ -5,10 +5,17 @@ import type { RollupError } from 'rollup'
 import { stripLiteral } from 'strip-literal'
 import type { ResolvedConfig } from '../config'
 import type { Plugin } from '../plugin'
-import { cleanUrl, injectQuery, normalizePath, parseRequest } from '../utils'
+import {
+  cleanUrl,
+  injectQuery,
+  normalizePath,
+  parseRequest,
+  transformResult
+} from '../utils'
 import type { WorkerType } from './worker'
 import { WORKER_FILE_ID, workerFileToUrl } from './worker'
 import { fileToUrl } from './asset'
+import { registerWorkersSource } from './optimizedDeps'
 
 const ignoreFlagRE = /\/\*\s*@vite-ignore\s*\*\//
 
@@ -113,8 +120,10 @@ export function workerImportMetaUrlPlugin(config: ResolvedConfig): Plugin {
           const file = normalizePath(
             path.resolve(path.dirname(id), rawUrl.slice(1, -1))
           )
+
           let url: string
           if (isBuild) {
+            registerWorkersSource(config, id)
             url = await workerFileToUrl(config, file, query)
           } else {
             url = await fileToUrl(cleanUrl(file), config, this)
@@ -127,10 +136,7 @@ export function workerImportMetaUrlPlugin(config: ResolvedConfig): Plugin {
         }
 
         if (s) {
-          return {
-            code: s.toString(),
-            map: config.build.sourcemap ? s.generateMap({ hires: true }) : null
-          }
+          return transformResult(s, id, config)
         }
 
         return null