+++ packages/vite/src/node/server/middlewares/transform.ts
@@ -34,7 +34,7 @@ import {
   ERR_OPTIMIZE_DEPS_PROCESSING_ERROR,
   ERR_OUTDATED_OPTIMIZED_DEP
 } from '../../plugins/optimizedDeps'
-import { createIsOptimizedDepUrl } from '../../optimizer'
+import { getDepsOptimizer } from '../../optimizer'
 
 const debugCache = createDebugger('vite:cache')
 const isDebug = !!process.env.DEBUG
@@ -48,8 +48,6 @@ export function transformMiddleware(
     config: { root, logger },
     moduleGraph
   } = server
-
-  const isOptimizedDepUrl = createIsOptimizedDepUrl(server.config)
 
   // Keep the named function. The name is visible in debug logs via `DEBUG=connect:dispatcher ...`
   return async function viteTransformMiddleware(req, res, next) {
@@ -73,7 +71,7 @@ export function transformMiddleware(
       const isSourceMap = withoutQuery.endsWith('.map')
       // since we generate source map references, handle those requests here
       if (isSourceMap) {
-        if (isOptimizedDepUrl(url)) {
+        if (getDepsOptimizer(server.config)?.isOptimizedDepUrl(url)) {
           // If the browser is requesting a source map for an optimized dep, it
           // means that the dependency has already been pre-bundled and loaded
           const mapFile = url.startsWith(FS_PREFIX)
@@ -124,14 +122,27 @@ export function transformMiddleware(
         const publicPath = `${publicDir.slice(rootDir.length)}/`
         // warn explicit public paths
         if (url.startsWith(publicPath)) {
-          logger.warn(
-            colors.yellow(
+          let warning: string
+
+          if (isImportRequest(url)) {
+            const rawUrl = removeImportQuery(url)
+
+            warning =
+              'Assets in public cannot be imported from JavaScript.\n' +
+              `Instead of ${colors.cyan(
+                rawUrl
+              )}, put the file in the src directory, and use ${colors.cyan(
+                rawUrl.replace(publicPath, '/src/')
+              )} instead.`
+          } else {
+            warning =
               `files in the public directory are served at the root path.\n` +
-                `Instead of ${colors.cyan(url)}, use ${colors.cyan(
-                  url.replace(publicPath, '/')
-                )}.`
-            )
-          )
+              `Instead of ${colors.cyan(url)}, use ${colors.cyan(
+                url.replace(publicPath, '/')
+              )}.`
+          }
+
+          logger.warn(colors.yellow(warning))
         }
       }
 
@@ -175,7 +186,9 @@ export function transformMiddleware(
         })
         if (result) {
           const type = isDirectCSSRequest(url) ? 'css' : 'js'
-          const isDep = DEP_VERSION_RE.test(url) || isOptimizedDepUrl(url)
+          const isDep =
+            DEP_VERSION_RE.test(url) ||
+            getDepsOptimizer(server.config)?.isOptimizedDepUrl(url)
           return send(req, res, result.code, type, {
             etag: result.etag,
             // allow browser to cache npm deps!